requirements.txt
# requirements.txt

python-telegram-bot==20.7
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
apscheduler==3.10.4
dotenv